// src/app/model/userReq.ts
/**
 * Turtle Nest Monitoring Persistence DB API
 * This projects establish the connection with the database, we don't have any logicimplemented in this project, since we only use this to save, consult and update information
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface UserReq { 
    id?: number;
    name?: string;
    lastName?: string;
    isActive?: boolean;
    title?: string;
    phone?: string;
    email?: string;
    password?: string;
    tempPassword?: string;
    isAdministrator?: boolean;
}