// src/app/model/sensorsReq.ts
/**
 * Turtle Nest Monitoring Persistence DB API
 * This projects establish the connection with the database, we don't have any logicimplemented in this project, since we only use this to save, consult and update information
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { OrganizationsReq } from './organizationsReq';

export interface SensorsReq { 
    id?: number;
    assignedID?: string;
    softwareVersion?: string;
    hardwareVersion?: string;
    components?: string;
    temperatureError?: number;
    moistureMinError?: number;
    moistureMaxError?: number;
    xerror?: number;
    yerror?: number;
    zerror?: number;
    isInUse?: boolean;
    dateManufactured?: Date;
    organization?: OrganizationsReq;
    nrSumar?: number;
    timing?: number;
}