// src/app/model/organizationsReq.ts
/**
 * Turtle Nest Monitoring Persistence DB API
 * This projects establish the connection with the database, we don't have any logicimplemented in this project, since we only use this to save, consult and update information
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { UserReq } from './userReq';

export interface OrganizationsReq { 
    id?: number;
    name?: string;
    dateActive?: Date;
    dateInactive?: Date;
    isActive?: boolean;
    contact?: UserReq;
    members?: Array<UserReq>;
}