// src/app/model/nestsReq.ts
/**
 * Turtle Nest Monitoring Persistence DB API
 * This projects establish the connection with the database, we don't have any logicimplemented in this project, since we only use this to save, consult and update information
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { OrganizationsReq } from './organizationsReq';
import { SensorsReq } from './sensorsReq';
import { UserReq } from './userReq';

/**
 * Nest object that needs to be added
 */
export interface NestsReq { 
    id?: number;
    assignedID?: string;
    latitude?: string;
    longitude?: string;
    sensor?: SensorsReq;
    organization?: OrganizationsReq;
    responsible?: UserReq;
    isActive?: boolean;
    eggsQuantity?: number;
    hatchlingsQuantity?: number;
    depredatedEggsQuantity?: number;
    infertileEggs?: number;
    zone?: string;
    layingDate?: Date;
    collectionDate?: Date;
    monitoringStartDate?: Date;
    monitoringEndDate?: Date;
    firstHatchingRecordedDate?: Date;
    lastHatchingRecordedDate?: Date;
    nestCleaningDate?: Date;
    nestBehavior?: string;
}